refactor to share HTTP logic
