+++ app/models/core_data_connector/item.rb
@@ -1,5 +1,3 @@
-require 'typhoeus'
-
 module CoreDataConnector
   class Item < ApplicationRecord
     # Includes
